+++ R/gg_bar.R
@@ -322,8 +322,9 @@ gg_bar <- function(data,
 #' @param col_labels_dp For numeric colour variables and where col_labels equals NULL, the number of decimal places. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
-#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "quantile".
+#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
+#' @param col_pretty_n For a numeric colour variable, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 6. 
 #' @param col_rev TRUE or FALSE of whether the colour scale is reversed. Defaults to FALSE. Defaults to FALSE.
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param col_title_wrap Number of characters to wrap the colour title to. Defaults to 25. Not applicable where mobile equals TRUE.
@@ -398,6 +399,7 @@ gg_bar_col <- function(data,
                        col_legend_nrow = NULL,
                        col_method = NULL,
                        col_na = TRUE,
+                       col_pretty_n = 6,
                        col_rev = FALSE,
                        col_title = NULL,
                        col_title_wrap = 25,
@@ -482,7 +484,7 @@ gg_bar_col <- function(data,
   
   if (is.null(col_method)) {
     if (!is.numeric(col_var_vctr)) col_method <- "category"
-    else if (is.numeric(col_var_vctr)) col_method <- "quantile"
+    else if (is.numeric(col_var_vctr)) col_method <- "bin"
   }
   
   if(col_method %in% c("quantile", "bin")) {
@@ -497,7 +499,7 @@ gg_bar_col <- function(data,
       if(is.null(col_labels_dp)) col_labels_dp <- 1
     }
     else if (col_method == "bin") {
-      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr)
+      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr, col_pretty_n)
       else({
         if (!(dplyr::first(col_cuts) %in% c(0, -Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
         if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")
@@ -1031,7 +1033,8 @@ gg_bar_facet <- function(data,
 #' @param col_labels_dp For numeric colour variables and where col_labels equals NULL, the number of decimal places. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
-#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "quantile".
+#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
+#' @param col_pretty_n For a numeric colour variable, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 6. 
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
 #' @param col_rev TRUE or FALSE of whether the colour scale is reversed. Defaults to FALSE. Defaults to FALSE.
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
@@ -1108,6 +1111,7 @@ gg_bar_col_facet <- function(data,
                              col_legend_nrow = NULL,
                              col_method = NULL,
                              col_na = TRUE,
+                             col_pretty_n = 6,
                              col_rev = FALSE,
                              col_title = NULL,
                              col_title_wrap = 25,
@@ -1209,7 +1213,7 @@ gg_bar_col_facet <- function(data,
   
   if (is.null(col_method)) {
     if (!is.numeric(col_var_vctr)) col_method <- "category"
-    else if (is.numeric(col_var_vctr)) col_method <- "quantile"
+    else if (is.numeric(col_var_vctr)) col_method <- "bin"
   }
   
   if(col_method %in% c("quantile", "bin")) {
@@ -1224,7 +1228,7 @@ gg_bar_col_facet <- function(data,
       if(is.null(col_labels_dp)) col_labels_dp <- 1
     }
     else if (col_method == "bin") {
-      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr)
+      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr, col_pretty_n)
       else({
         if (!(dplyr::first(col_cuts) %in% c(0, -Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
         if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")