+++ R/gg_hbar.R
@@ -327,8 +327,9 @@ gg_hbar <- function(data,
 #' @param col_labels_dp For numeric colour variables and where col_labels equals NULL, the number of decimal places. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
-#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "quantile".
+#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
+#' @param col_pretty_n For a numeric colour variable, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 6. 
 #' @param col_rev TRUE or FALSE of whether the colour scale is reversed. Defaults to FALSE. Defaults to FALSE.
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param col_title_wrap Number of characters to wrap the colour title to. Defaults to 25. Not applicable where mobile equals TRUE.
@@ -404,6 +405,7 @@ gg_hbar_col <- function(data,
                         col_legend_nrow = NULL,
                         col_method = NULL,
                         col_na = TRUE,
+                        col_pretty_n = 6,
                         col_rev = FALSE,
                         col_title = NULL,
                         col_title_wrap = 25,
@@ -492,7 +494,7 @@ gg_hbar_col <- function(data,
   else col_n <- length(unique(col_var_vctr))
   if (is.null(col_method)) {
     if (!is.numeric(col_var_vctr)) col_method <- "category"
-    else if (is.numeric(col_var_vctr)) col_method <- "quantile"
+    else if (is.numeric(col_var_vctr)) col_method <- "bin"
   }
   
   if(col_method %in% c("quantile", "bin")) {
@@ -507,7 +509,7 @@ gg_hbar_col <- function(data,
       if(is.null(col_labels_dp)) col_labels_dp <- 1
     }
     else if (col_method == "bin") {
-      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr)
+      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr, col_pretty_n)
       else({
         if (!(dplyr::first(col_cuts) %in% c(0, -Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
         if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")
@@ -1060,8 +1062,9 @@ gg_hbar_facet <- function(data,
 #' @param col_labels_dp For numeric colour variables and where col_labels equals NULL, the number of decimal places. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
-#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "quantile".
+#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
+#' @param col_pretty_n For a numeric colour variable, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 6. 
 #' @param col_rev TRUE or FALSE of whether the colour scale is reversed. Defaults to FALSE. Defaults to FALSE.
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param col_title_wrap Number of characters to wrap the colour title to. Defaults to 25. Not applicable where mobile equals TRUE.
@@ -1137,6 +1140,7 @@ gg_hbar_col_facet <- function(data,
                               col_legend_nrow = NULL,
                               col_method = NULL,
                               col_na = TRUE,
+                              col_pretty_n = 6,
                               col_rev = FALSE,
                               col_title = NULL,
                               col_title_wrap = 25,
@@ -1238,7 +1242,7 @@ gg_hbar_col_facet <- function(data,
   
   if (is.null(col_method)) {
     if (!is.numeric(col_var_vctr)) col_method <- "category"
-    else if (is.numeric(col_var_vctr)) col_method <- "quantile"
+    else if (is.numeric(col_var_vctr)) col_method <- "bin"
   }
   
   if(col_method %in% c("quantile", "bin")) {
@@ -1253,7 +1257,7 @@ gg_hbar_col_facet <- function(data,
       if(is.null(col_labels_dp)) col_labels_dp <- 1
     }
     else if (col_method == "bin") {
-      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr)
+      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr, col_pretty_n)
       else({
         if (!(dplyr::first(col_cuts) %in% c(0, -Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
         if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")