+++ R/gg_point.R
@@ -315,8 +315,9 @@ gg_point <- function(data,
 #' @param col_labels_dp For numeric colour variables and where col_labels equals NULL, the number of decimal places. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
-#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "quantile".
+#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
+#' @param col_pretty_n For a numeric colour variable, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 6. 
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param col_title_wrap Number of characters to wrap the colour title to. Defaults to 25. Not applicable where mobile equals TRUE.
 #' @param caption Caption title string. 
@@ -378,6 +379,7 @@ gg_point_col <- function(data,
                          col_legend_nrow = NULL,
                          col_method = NULL,
                          col_na = TRUE,
+                         col_pretty_n = 6,
                          font_family = "",
                          font_size_title = NULL,
                          font_size_body = NULL,
@@ -449,7 +451,7 @@ gg_point_col <- function(data,
   
   if (is.null(col_method)) {
     if (!is.numeric(col_var_vctr)) col_method <- "category"
-    else if (is.numeric(col_var_vctr)) col_method <- "quantile"
+    else if (is.numeric(col_var_vctr)) col_method <- "bin"
   }
   
   if(col_method %in% c("quantile", "bin")) {
@@ -464,7 +466,7 @@ gg_point_col <- function(data,
       if(is.null(col_labels_dp)) col_labels_dp <- 1
     }
     else if (col_method == "bin") {
-      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr)
+      if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr, col_pretty_n)
       else({
         if (!(dplyr::first(col_cuts) %in% c(0, -Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
         if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")
@@ -975,8 +977,9 @@ gg_point_facet <- function(data,
 #' @param col_labels_dp For numeric colour variables and where col_labels equals NULL, the number of decimal places. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
-#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "quantile".
+#' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
+#' @param col_pretty_n For a numeric colour variable, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 6. 
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param col_title_wrap Number of characters to wrap the colour title to. Defaults to 25. 
 #' @param facet_labels As per the ggplot2 labeller argument within the ggplot facet_wrap function. If NULL, defaults to ggplot2::as_labeller(stringr::str_to_sentence). Use facet_labels = ggplot2::label_value to turn off default sentence case transformation.
@@ -1047,6 +1050,7 @@ gg_point_col_facet <-
            col_legend_nrow = NULL,
            col_method = NULL,
            col_na = TRUE,
+           col_pretty_n = 6,
            col_title = NULL,
            col_title_wrap = 25,
            facet_labels = NULL,
@@ -1133,7 +1137,7 @@ gg_point_col_facet <-
     
     if (is.null(col_method)) {
       if (!is.numeric(col_var_vctr)) col_method <- "category"
-      else if (is.numeric(col_var_vctr)) col_method <- "quantile"
+      else if (is.numeric(col_var_vctr)) col_method <- "bin"
     }
     
     if(col_method %in% c("quantile", "bin")) {
@@ -1148,7 +1152,7 @@ gg_point_col_facet <-
         if(is.null(col_labels_dp)) col_labels_dp <- 1
       }
       else if (col_method == "bin") {
-        if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr)
+        if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr, col_pretty_n)
         else({
           if (!(dplyr::first(col_cuts) %in% c(0,-Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
           if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")